+++ tests/plugins/test_base_plugin.py
@@ -4,8 +4,8 @@
 # SPDX-License-Identifier: GPL-3.0-only
 #
 # pydidas is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License version 3 as published by
-# the Free Software Foundation.
+# it under the terms of the GNU General Public License version 3 as
+# published by the Free Software Foundation.
 #
 # Pydidas is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -18,10 +18,10 @@
 """Unit tests for pydidas modules."""
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2021-2022, Malte Storm, Helmholtz-Zentrum Hereon"
-__license__ = "GPL-3.0"
+__copyright__ = "Copyright 2023, Helmholtz-Zentrum Hereon"
+__license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
-__status__ = "Development"
+__status__ = "Production"
 
 
 import unittest
@@ -32,7 +32,7 @@ import pickle
 
 import numpy as np
 
-from pydidas.core import Parameter, get_generic_parameter, utils
+from pydidas.core import Parameter, get_generic_parameter, utils, UserConfigError
 from pydidas.core.constants import BASE_PLUGIN
 from pydidas.core.utils import rebin2d
 from pydidas.unittest_objects import create_plugin_class
@@ -320,7 +320,14 @@ class TestBasePlugin(unittest.TestCase):
     def test_input_shape_setter__wrong_type(self):
         _shape = 123
         plugin = create_plugin_class(BASE_PLUGIN)()
-        with self.assertRaises(TypeError):
+        with self.assertRaises(UserConfigError):
+            plugin.input_shape = _shape
+
+    def test_input_shape_setter__input_data_dim_None(self):
+        _shape = (123, 456)
+        plugin = create_plugin_class(BASE_PLUGIN)()
+        plugin.input_data_dim = None
+        with self.assertRaises(UserConfigError):
             plugin.input_shape = _shape
 
     def test_input_shape_setter__input_data_dim_neg(self):
@@ -333,7 +340,7 @@ class TestBasePlugin(unittest.TestCase):
         _shape = (123, 534, 245)
         plugin = create_plugin_class(BASE_PLUGIN)()
         plugin.input_data_dim = 2
-        with self.assertRaises(ValueError):
+        with self.assertRaises(UserConfigError):
             plugin.input_shape = _shape
 
     def test_input_shape_setter__input_data_dim_correct(self):